Added Schedule Functionality
Schedule now displays on mainactivity for the appropriate user.
Starting with the current day, the schedule displays the month, day of
month, day of week, and first shift hours of each day.

Note: Currently only displays hours of first shift each day and displays
it in military time.
This will be ironed out.

Known bug: Signing out and right back in may result in blank screen in
mainactivity.